+++ src/senaite/app/supermodel/model.py
@@ -222,6 +222,12 @@ class SuperModel(object):
     def get_field_value(self, name, default=None):
         """Returns the value for the given name and current instance
         """
+        # These are special "fields" that are widely accessed in lowercase,
+        # but we always need to rely on the capitalized function
+        if name.lower() in ["title", "description"]:
+            func = getattr(self.instance, name.capitalize(), None)
+            return func() if func else default
+
         # always give priority to getters regardless of type
         accessor_name = "get{}".format(name)
         accessor = getattr(self.instance, accessor_name, _marker)